+++ client/src/components/MobileNavigation.tsx
@@ -1,10 +1,13 @@
+// client/src/components/MobileNavigation.tsx
+// Responsabilité unique : Navigation principale mobile avec orchestration
+
 import React, { useState } from 'react';
 import { Link, useLocation } from 'react-router-dom';
-import { HomeIcon, SearchIcon, UserIcon, ShoppingCartIcon, HeartIcon, StoreIcon, BarsIcon, TimesIcon } from '../utils/icons';
+import { HomeIcon, SearchIcon, UserIcon, StoreIcon, BarsIcon, TimesIcon } from '../utils/icons';
 import { useAuth } from '../context/AuthContext';
-import { useCartStore } from '../stores/useCartStore';
-import { useWishlistStore } from '../stores/useWishlistStore';
 import { TouchButton } from './TouchButton';
+import FloatingActions from './FloatingActions';
+import MobileMenuOverlay from './MobileMenuOverlay';
 
 interface MobileNavigationProps {
   className?: string;
@@ -14,9 +17,6 @@ export const MobileNavigation: React.FC<MobileNavigationProps> = ({ className = 
   const { isAuthenticated } = useAuth();
   const [isMenuOpen, setIsMenuOpen] = useState(false);
   const location = useLocation();
-
-  const cartItemCount = useCartStore((state) => state.cartItems.length);
-  const wishlistItemCount = useWishlistStore((state) => state.wishlistItems.length);
 
   const isActive = (path: string) => location.pathname === path;
 
@@ -80,104 +80,11 @@ export const MobileNavigation: React.FC<MobileNavigationProps> = ({ className = 
         </div>
 
         {/* Quick Actions Floating */}
-        {isAuthenticated && (
-          <div className="absolute -top-16 right-4 flex space-x-2">
-            <Link
-              to="/cart"
-              className={`
-                relative touch-target bg-special text-primary rounded-full p-3 shadow-lg
-                transition-transform duration-200 hover:scale-110
-                ${isActive('/cart') ? 'scale-110' : ''}
-              `}
-            >
-              <ShoppingCartIcon size={20} />
-              {cartItemCount > 0 && (
-                <span className="absolute -top-1 -right-1 bg-red-500 text-white text-xs rounded-full h-5 w-5 flex items-center justify-center font-bold">
-                  {cartItemCount > 9 ? '9+' : cartItemCount}
-                </span>
-              )}
-            </Link>
-
-            <Link
-              to="/wishlist"
-              className={`
-                relative touch-target bg-primary text-text rounded-full p-3 shadow-lg
-                transition-transform duration-200 hover:scale-110
-                ${isActive('/wishlist') ? 'scale-110' : ''}
-              `}
-            >
-              <HeartIcon size={20} />
-              {wishlistItemCount > 0 && (
-                <span className="absolute -top-1 -right-1 bg-special text-primary text-xs rounded-full h-5 w-5 flex items-center justify-center font-bold">
-                  {wishlistItemCount > 9 ? '9+' : wishlistItemCount}
-                </span>
-              )}
-            </Link>
-          </div>
-        )}
+        {isAuthenticated && <FloatingActions />}
       </nav>
 
       {/* Slide-up Menu Overlay */}
-      {isMenuOpen && (
-        <div
-          className="fixed inset-0 z-40 bg-black/50 backdrop-blur-sm md:hidden"
-          onClick={closeMenu}
-        >
-          <div
-            className="absolute bottom-0 left-0 right-0 bg-secondary rounded-t-2xl p-6 pb-24 transform transition-transform duration-300 ease-out"
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div className="flex flex-col space-y-4">
-              <div className="text-center mb-4">
-                <h3 className="text-lg font-display text-text">Menu</h3>
-                <div className="w-12 h-1 bg-primary rounded mx-auto mt-2"></div>
-              </div>
-
-              <div className="grid grid-cols-2 gap-4">
-                {[
-                  { path: '/about', label: 'À propos', icon: '🌟' },
-                  { path: '/contact', label: 'Contact', icon: '📞' },
-                  { path: '/faq', label: 'FAQ', icon: '❓' },
-                  { path: '/legal', label: 'Mentions légales', icon: '📋' },
-                ].map((item) => (
-                  <Link
-                    key={item.path}
-                    to={item.path}
-                    onClick={closeMenu}
-                    className="mobile-card bg-primary/10 hover:bg-primary/20 transition-colors duration-200 text-center"
-                  >
-                    <div className="text-2xl mb-2">{item.icon}</div>
-                    <div className="text-sm font-medium text-text">{item.label}</div>
-                  </Link>
-                ))}
-              </div>
-
-              {isAuthenticated && (
-                <div className="border-t border-primary/20 pt-4 mt-6">
-                  <div className="flex justify-around">
-                    <Link
-                      to="/orders"
-                      onClick={closeMenu}
-                      className="flex items-center space-x-2 text-text hover:text-special transition-colors"
-                    >
-                      <span>📦</span>
-                      <span>Mes commandes</span>
-                    </Link>
-                    <Link
-                      to="/settings"
-                      onClick={closeMenu}
-                      className="flex items-center space-x-2 text-text hover:text-special transition-colors"
-                    >
-                      <span>⚙️</span>
-                      <span>Paramètres</span>
-                    </Link>
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
+      <MobileMenuOverlay isOpen={isMenuOpen} onClose={closeMenu} />
     </>
   );
 };
